Generalize the use of size_t in TimeIntegration

See merge request nektar/nektar!1555

+++ library/LibUtilities/TimeIntegration/ExplicitTimeIntegrationSchemeSDC.h
@@ -49,7 +49,7 @@ namespace LibUtilities
 class ExplicitTimeIntegrationSchemeSDC : public TimeIntegrationSchemeSDC
 {
 public:
-    ExplicitTimeIntegrationSchemeSDC(std::string variant, unsigned int order,
+    ExplicitTimeIntegrationSchemeSDC(std::string variant, size_t order,
                                      std::vector<NekDouble> freeParams)
         : TimeIntegrationSchemeSDC(variant, order, freeParams)
     {
@@ -58,8 +58,7 @@ public:
     }
 
     static TimeIntegrationSchemeSharedPtr create(
-        std::string variant, unsigned int order,
-        std::vector<NekDouble> freeParams)
+        std::string variant, size_t order, std::vector<NekDouble> freeParams)
     {
         TimeIntegrationSchemeSharedPtr p =
             MemoryManager<ExplicitTimeIntegrationSchemeSDC>::AllocateSharedPtr(
