+++ library/LibUtilities/TimeIntegration/IMEXTimeIntegrationSchemeSDC.h
@@ -49,7 +49,7 @@ namespace LibUtilities
 class IMEXTimeIntegrationSchemeSDC : public TimeIntegrationSchemeSDC
 {
 public:
-    IMEXTimeIntegrationSchemeSDC(std::string variant, unsigned int order,
+    IMEXTimeIntegrationSchemeSDC(std::string variant, size_t order,
                                  std::vector<NekDouble> freeParams)
         : TimeIntegrationSchemeSDC(variant, order, freeParams)
     {
@@ -69,8 +69,7 @@ public:
     }
 
     static TimeIntegrationSchemeSharedPtr create(
-        std::string variant, unsigned int order,
-        std::vector<NekDouble> freeParams)
+        std::string variant, size_t order, std::vector<NekDouble> freeParams)
     {
         TimeIntegrationSchemeSharedPtr p =
             MemoryManager<IMEXTimeIntegrationSchemeSDC>::AllocateSharedPtr(
